Refactor backend to standard FastAPI structure
This commit refactors the entire backend application to follow a standard, cleaner FastAPI project structure.

Key changes include:
- A new `app` directory containing all application logic.
- Separation of concerns into `api`, `core`, `crud`, `db`, `models`, `schemas`, and `services` modules.
- Introduction of a generic `CRUDBase` for reusable database operations.
- Updated `Dockerfile` to reflect the new project structure.
- Moved utility scripts (`build_vector_store.py`, `seed.py`) into `app/scripts`.
- Resolved pathing issues for database and other resources.
