+++ backend/app/api/deps.py
@@ -1,0 +1,9 @@
+from app.db.session import SessionLocal
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
