+++ backend/app/models/__init__.py
@@ -1,0 +1,2 @@
+from .product import Product
+from .order import Order
