+++ backend/app/models/order.py
@@ -1,0 +1,15 @@
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey
+from sqlalchemy.orm import relationship
+from sqlalchemy.sql import func
+
+from app.db.base_class import Base
+
+class Order(Base):
+    id = Column(Integer, primary_key=True, index=True)
+    user_id = Column(String, index=True, default="user123")
+    product_id = Column(Integer, ForeignKey("products.id"))
+    quantity = Column(Integer, default=1)
+    status = Column(String, default="Pending")
+    created_at = Column(DateTime(timezone=True), server_default=func.now())
+
+    product = relationship("Product", back_populates="orders")
