+++ backend/app/schemas/__init__.py
@@ -1,0 +1,3 @@
+from .product import Product, ProductBase, StockStatus
+from .order import Order, OrderBase, OrderStatus
+from .knowledge import KnowledgeAnswer, QuestionRequest
