+++ backend/app/schemas/knowledge.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+class KnowledgeAnswer(BaseModel):
+    question: str
+    answer: str
+    source: str
+
+class QuestionRequest(BaseModel):
+    question: str
