+++ backend/app/scripts/seed.py
@@ -1,21 +1,24 @@
 from sqlalchemy.orm import Session
-from .database import SessionLocal, engine
-from . import models
+
+from app.db.session import SessionLocal, engine
+from app.db.base_class import Base
+from app.models.order import Order
+from app.models.product import Product
 
 # Create tables
-models.Base.metadata.create_all(bind=engine)
+Base.metadata.create_all(bind=engine)
 
 def seed_data():
     db: Session = SessionLocal()
 
     # Clear existing data
-    db.query(models.Order).delete()
-    db.query(models.Product).delete()
+    db.query(Order).delete()
+    db.query(Product).delete()
     db.commit()
 
     # --- Create Products ---
     products_to_add = [
-        models.Product(
+        Product(
             name="iPhone 15 Pro",
             description="The latest and greatest iPhone with the A17 Pro chip.",
             price=999.99,
@@ -23,7 +26,7 @@ def seed_data():
             category="Electronics",
             specs={"color": "Titanium Blue", "storage": "256GB"},
         ),
-        models.Product(
+        Product(
             name="MacBook Pro 14-inch",
             description="A powerful laptop for professionals with the M3 chip.",
             price=1999.00,
@@ -31,7 +34,7 @@ def seed_data():
             category="Electronics",
             specs={"chip": "M3 Pro", "ram": "18GB", "storage": "512GB SSD"},
         ),
-        models.Product(
+        Product(
             name="Sony WH-1000XM5 Headphones",
             description="Industry-leading noise-canceling headphones.",
             price=399.00,
@@ -39,7 +42,7 @@ def seed_data():
             category="Accessories",
             specs={"color": "Black", "type": "Over-ear"},
         ),
-        models.Product(
+        Product(
             name="Organic Cotton T-Shirt",
             description="A comfortable and stylish t-shirt made from 100% organic cotton.",
             price=25.50,
@@ -47,7 +50,7 @@ def seed_data():
             category="Apparel",
             specs={"color": "White", "size": "M"},
         ),
-         models.Product(
+         Product(
             name="iPhone 15 Pro",
             description="The latest and greatest iPhone with the A17 Pro chip.",
             price=1099.99,
@@ -60,24 +63,24 @@ def seed_data():
     db.commit()
 
     # --- Create Orders ---
-    iphone_prod = db.query(models.Product).filter(models.Product.name == "iPhone 15 Pro").first()
-    macbook_prod = db.query(models.Product).filter(models.Product.name == "MacBook Pro 14-inch").first()
+    iphone_prod = db.query(Product).filter(Product.name == "iPhone 15 Pro").first()
+    macbook_prod = db.query(Product).filter(Product.name == "MacBook Pro 14-inch").first()
 
     if iphone_prod and macbook_prod:
         orders_to_add = [
-            models.Order(
+            Order(
                 product_id=iphone_prod.id,
                 quantity=1,
                 status="Shipped",
                 user_id="user456"
             ),
-            models.Order(
+            Order(
                 product_id=macbook_prod.id,
                 quantity=1,
                 status="Delivered",
                 user_id="user789"
             ),
-            models.Order(
+            Order(
                 id=12345, # Specific order ID for testing
                 product_id=iphone_prod.id,
                 quantity=1,
