+++ backend/app/services/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the 'services' directory a Python package.
