+++ backend/app/services/embeddings.py
