+++ backend/app/services/knowledge_service.py
